Fixed compilation issues on cygwin:
1) added the missing __YICES_DLLSPEC__ hack in yices_smtcomp.c
   and yices_smt.c

2) fixed compilation warnings caused by issalpha, isdigit, isxdigit, etc,
   by adding a type coercion to (int)

3) fixed static_bin build by adding option -static-libgcc to the 
   STATIC variable in src/Makefile (check README for explanations).



git-svn-id: svn+ssh://rfs.csl.sri.com/svn/private/yices2/branches/new-terms@1699 0c041c21-16df-db11-9388-003048836090

+++ src/bv64_constants.c
@@ -249,7 +249,7 @@ int32_t bvconst64_set_from_hexa_string(uint64_t *a, uint32_t n, char *s) {
   do {
     c = *s;
     s ++;
-    if (isxdigit(c)) {
+    if (isxdigit((int) c)) {
       hex = hextoint(c);
       assert(0 <= hex && hex < 16);
       // set bits 4n-1 to 4n-4
